requirements.txt
streamlit
langgraph
langchain
openai
pydantic
spacy
charset-normalizer
pdfplumber 
python-docx